LayoutTests/imported/w3c/web-platform-tests/service-workers/service-worker/fetch-event-redirect.https-expected.txt


PASS initialize global state 
FAIL Non-navigation, manual redirect, cors mode Request redirected to same-origin without credentials should succeed opaqueredirect interception and response should not be redirected assert_equals: response.type expected "opaqueredirect" but got "basic"
FAIL Non-navigation, manual redirect, cors mode Request redirected to no-cors without credentials should succeed opaqueredirect interception and response should not be redirected assert_equals: response.type expected "opaqueredirect" but got "basic"
FAIL Non-navigation, manual redirect, cors mode Request redirected to cors without credentials should succeed opaqueredirect interception and response should not be redirected assert_equals: response.type expected "opaqueredirect" but got "basic"
FAIL Non-navigation, manual redirect, same-origin mode Request redirected to same-origin without credentials should succeed opaqueredirect interception and response should not be redirected assert_equals: response.type expected "opaqueredirect" but got "basic"
FAIL Non-navigation, manual redirect, same-origin mode Request redirected to no-cors without credentials should succeed opaqueredirect interception and response should not be redirected assert_equals: response.type expected "opaqueredirect" but got "basic"
FAIL Non-navigation, manual redirect, same-origin mode Request redirected to cors without credentials should succeed opaqueredirect interception and response should not be redirected assert_equals: response.type expected "opaqueredirect" but got "basic"
FAIL Non-navigation, manual redirect, no-cors mode Request redirected to same-origin without credentials should succeed opaqueredirect interception and response should not be redirected assert_equals: response.type expected "opaqueredirect" but got "basic"
FAIL Non-navigation, manual redirect, no-cors mode Request redirected to no-cors without credentials should succeed interception and response should not be redirected assert_equals: response.type expected "opaqueredirect" but got "basic"
FAIL Non-navigation, manual redirect, no-cors mode Request redirected to cors without credentials should succeed interception and response should not be redirected assert_equals: response.type expected "opaqueredirect" but got "basic"
FAIL Non-navigation, manual redirect, cors mode Request redirected to same-origin with credentials should succeed opaqueredirect interception and response should not be redirected assert_equals: response.type expected "opaqueredirect" but got "basic"
FAIL Non-navigation, manual redirect, cors mode Request redirected to no-cors with credentials should succeed opaqueredirect interception and response should not be redirected assert_equals: response.type expected "opaqueredirect" but got "basic"
FAIL Non-navigation, manual redirect, cors mode Request redirected to cors with credentials should succeed opaqueredirect interception and response should not be redirected assert_equals: response.type expected "opaqueredirect" but got "basic"
FAIL Non-navigation, manual redirect, same-origin mode Request redirected to same-origin with credentials should succeed opaqueredirect interception and response should not be redirected assert_equals: response.type expected "opaqueredirect" but got "basic"
FAIL Non-navigation, manual redirect, same-origin mode Request redirected to no-cors with credentials should succeed opaqueredirect interception and response should not be redirected assert_equals: response.type expected "opaqueredirect" but got "basic"
FAIL Non-navigation, manual redirect, same-origin mode Request redirected to cors with credentials should succeed opaqueredirect interception and response should not be redirected assert_equals: response.type expected "opaqueredirect" but got "basic"
FAIL Non-navigation, manual redirect, no-cors mode Request redirected to same-origin with credentials should succeed opaqueredirect interception and response should not be redirected assert_equals: response.type expected "opaqueredirect" but got "basic"
FAIL Non-navigation, manual redirect, no-cors mode Request redirected to no-cors with credentials should succeed interception and response should not be redirected assert_equals: response.type expected "opaqueredirect" but got "basic"
FAIL Non-navigation, manual redirect, no-cors mode Request redirected to cors with credentials should succeed interception and response should not be redirected assert_equals: response.type expected "opaqueredirect" but got "basic"
FAIL Non-navigation, follow redirect, cors mode Request redirected to same-origin without credentials should succeed interception and response should be redirected assert_equals: response.redirected expected true but got false
FAIL Non-navigation, follow redirect, cors mode Request redirected to no-cors without credentials should fail interception and response should not be redirected promise_test: Unhandled rejection with value: "assert_promise_rejects: Must fail to fetch: url=/nonav-follow-cors-redirects-to-nocors-nocreds?url=redirect.py%3FRedirect%3Dhttps%253A%252F%252F127.0.0.1%253A9443%252Fservice-workers%252Fservice-worker%252Fresources%252Fsuccess.py Promise did not reject."
FAIL Non-navigation, follow redirect, cors mode Request redirected to cors without credentials should succeed interception and response should be redirected assert_equals: response.type expected "cors" but got "basic"
FAIL Non-navigation, follow redirect, same-origin mode Request redirected to same-origin without credentials should succeed interception and response should be redirected assert_equals: response.redirected expected true but got false
FAIL Non-navigation, follow redirect, same-origin mode Request redirected to no-cors without credentials should fail interception and response should not be redirected promise_test: Unhandled rejection with value: "assert_promise_rejects: Must fail to fetch: url=/nonav-follow-sameorigin-redirects-to-nocors-nocreds?url=redirect.py%3FRedirect%3Dhttps%253A%252F%252F127.0.0.1%253A9443%252Fservice-workers%252Fservice-worker%252Fresources%252Fsuccess.py Promise did not reject."
FAIL Non-navigation, follow redirect, same-origin mode Request redirected to cors without credentials should fail interception and response should not be redirected promise_test: Unhandled rejection with value: "assert_promise_rejects: Must fail to fetch: url=/nonav-follow-sameorigin-redirects-to-cors-nocreds?url=redirect.py%3FRedirect%3Dhttps%253A%252F%252F127.0.0.1%253A9443%252Fservice-workers%252Fservice-worker%252Fresources%252Fsuccess.py%253FACAOrigin%253Dhttps%25253A%25252F%25252Flocalhost%25253A9443 Promise did not reject."
FAIL Non-navigation, follow redirect, no-cors mode Request redirected to same-origin without credentials should succeed interception and response should be redirected assert_equals: response.redirected expected true but got false
FAIL Non-navigation, follow redirect, no-cors mode Request redirected to no-cors without credentials should succeed interception and response should not be redirected assert_equals: response.type expected "opaque" but got "basic"
FAIL Non-navigation, follow redirect, no-cors mode Request redirected to cors without credentials should succeed interception and response should not be redirected assert_equals: response.type expected "opaque" but got "basic"
FAIL Non-navigation, follow redirect, cors mode Request redirected to same-origin with credentials should succeed interception and response should be redirected assert_equals: response.redirected expected true but got false
FAIL Non-navigation, follow redirect, cors mode Request redirected to no-cors with credentials should fail interception and response should not be redirected promise_test: Unhandled rejection with value: "assert_promise_rejects: Must fail to fetch: url=/nonav-follow-cors-redirects-to-nocors-creds?url=redirect.py%3FRedirect%3Dhttps%253A%252F%252Ffoo%253Abar%2540127.0.0.1%253A9443%252Fservice-workers%252Fservice-worker%252Fresources%252Fsuccess.py Promise did not reject."
FAIL Non-navigation, follow redirect, cors mode Request redirected to cors with credentials should fail interception and response should be redirected promise_test: Unhandled rejection with value: "assert_promise_rejects: Must fail to fetch: url=/nonav-follow-cors-redirects-to-cors-creds?url=redirect.py%3FRedirect%3Dhttps%253A%252F%252Ffoo%253Abar%2540127.0.0.1%253A9443%252Fservice-workers%252Fservice-worker%252Fresources%252Fsuccess.py%253FACAOrigin%253Dhttps%25253A%25252F%25252Flocalhost%25253A9443 Promise did not reject."
FAIL Non-navigation, follow redirect, same-origin mode Request redirected to same-origin with credentials should succeed interception and response should be redirected assert_equals: response.redirected expected true but got false
FAIL Non-navigation, follow redirect, same-origin mode Request redirected to no-cors with credentials should fail interception and response should not be redirected promise_test: Unhandled rejection with value: "assert_promise_rejects: Must fail to fetch: url=/nonav-follow-sameorigin-redirects-to-nocors-creds?url=redirect.py%3FRedirect%3Dhttps%253A%252F%252Ffoo%253Abar%2540127.0.0.1%253A9443%252Fservice-workers%252Fservice-worker%252Fresources%252Fsuccess.py Promise did not reject."
FAIL Non-navigation, follow redirect, same-origin mode Request redirected to cors with credentials should fail interception and response should not be redirected promise_test: Unhandled rejection with value: "assert_promise_rejects: Must fail to fetch: url=/nonav-follow-sameorigin-redirects-to-cors-creds?url=redirect.py%3FRedirect%3Dhttps%253A%252F%252Ffoo%253Abar%2540127.0.0.1%253A9443%252Fservice-workers%252Fservice-worker%252Fresources%252Fsuccess.py%253FACAOrigin%253Dhttps%25253A%25252F%25252Flocalhost%25253A9443 Promise did not reject."
FAIL Non-navigation, follow redirect, no-cors mode Request redirected to same-origin with credentials should succeed interception and response should be redirected assert_equals: response.redirected expected true but got false
FAIL Non-navigation, follow redirect, no-cors mode Request redirected to no-cors with credentials should succeed interception and response should not be redirected assert_equals: response.type expected "opaque" but got "basic"
FAIL Non-navigation, follow redirect, no-cors mode Request redirected to cors with credentials should succeed interception and response should not be redirected assert_equals: response.type expected "opaque" but got "basic"
FAIL Non-navigation, error redirect, cors mode Request redirected to same-origin without credentials should fail interception and response should not be redirected promise_test: Unhandled rejection with value: "assert_promise_rejects: Must fail to fetch: url=/nonav-error-cors-redirects-to-sameorigin-nocreds?url=redirect.py%3FRedirect%3Dhttps%253A%252F%252Flocalhost%253A9443%252Fservice-workers%252Fservice-worker%252Fresources%252Fsuccess.py Promise did not reject."
FAIL Non-navigation, error redirect, cors mode Request redirected to no-cors without credentials should fail interception and response should not be redirected promise_test: Unhandled rejection with value: "assert_promise_rejects: Must fail to fetch: url=/nonav-error-cors-redirects-to-nocors-nocreds?url=redirect.py%3FRedirect%3Dhttps%253A%252F%252F127.0.0.1%253A9443%252Fservice-workers%252Fservice-worker%252Fresources%252Fsuccess.py Promise did not reject."
FAIL Non-navigation, error redirect, cors mode Request redirected to cors without credentials should fail interception and response should not be redirected promise_test: Unhandled rejection with value: "assert_promise_rejects: Must fail to fetch: url=/nonav-error-cors-redirects-to-cors-nocreds?url=redirect.py%3FRedirect%3Dhttps%253A%252F%252F127.0.0.1%253A9443%252Fservice-workers%252Fservice-worker%252Fresources%252Fsuccess.py%253FACAOrigin%253Dhttps%25253A%25252F%25252Flocalhost%25253A9443 Promise did not reject."
FAIL Non-navigation, error redirect, same-origin mode Request redirected to same-origin without credentials should fail interception and response should not be redirected promise_test: Unhandled rejection with value: "assert_promise_rejects: Must fail to fetch: url=/nonav-error-sameorigin-redirects-to-sameorigin-nocreds?url=redirect.py%3FRedirect%3Dhttps%253A%252F%252Flocalhost%253A9443%252Fservice-workers%252Fservice-worker%252Fresources%252Fsuccess.py Promise did not reject."
FAIL Non-navigation, error redirect, same-origin mode Request redirected to no-cors without credentials should fail interception and response should not be redirected promise_test: Unhandled rejection with value: "assert_promise_rejects: Must fail to fetch: url=/nonav-error-sameorigin-redirects-to-nocors-nocreds?url=redirect.py%3FRedirect%3Dhttps%253A%252F%252F127.0.0.1%253A9443%252Fservice-workers%252Fservice-worker%252Fresources%252Fsuccess.py Promise did not reject."
FAIL Non-navigation, error redirect, same-origin mode Request redirected to cors without credentials should fail interception and response should not be redirected promise_test: Unhandled rejection with value: "assert_promise_rejects: Must fail to fetch: url=/nonav-error-sameorigin-redirects-to-cors-nocreds?url=redirect.py%3FRedirect%3Dhttps%253A%252F%252F127.0.0.1%253A9443%252Fservice-workers%252Fservice-worker%252Fresources%252Fsuccess.py%253FACAOrigin%253Dhttps%25253A%25252F%25252Flocalhost%25253A9443 Promise did not reject."
FAIL Non-navigation, error redirect, no-cors mode Request redirected to same-origin without credentials should fail interception and response should not be redirected promise_test: Unhandled rejection with value: "assert_promise_rejects: Must fail to fetch: url=/nonav-error-nocors-redirects-to-sameorigin-nocreds?url=redirect.py%3FRedirect%3Dhttps%253A%252F%252Flocalhost%253A9443%252Fservice-workers%252Fservice-worker%252Fresources%252Fsuccess.py Promise did not reject."
FAIL Non-navigation, error redirect, no-cors mode Request redirected to no-cors without credentials should fail interception and response should not be redirected promise_test: Unhandled rejection with value: "assert_promise_rejects: Must fail to fetch: url=/nonav-error-nocors-redirects-to-nocors-nocreds?url=redirect.py%3FRedirect%3Dhttps%253A%252F%252F127.0.0.1%253A9443%252Fservice-workers%252Fservice-worker%252Fresources%252Fsuccess.py Promise did not reject."
FAIL Non-navigation, error redirect, no-cors mode Request redirected to cors without credentials should fail interception and response should not be redirected promise_test: Unhandled rejection with value: "assert_promise_rejects: Must fail to fetch: url=/nonav-error-nocors-redirects-to-cors-nocreds?url=redirect.py%3FRedirect%3Dhttps%253A%252F%252F127.0.0.1%253A9443%252Fservice-workers%252Fservice-worker%252Fresources%252Fsuccess.py%253FACAOrigin%253Dhttps%25253A%25252F%25252Flocalhost%25253A9443 Promise did not reject."
FAIL Non-navigation, error redirect, cors mode Request redirected to same-origin with credentials should fail interception and response should not be redirected promise_test: Unhandled rejection with value: "assert_promise_rejects: Must fail to fetch: url=/nonav-error-cors-redirects-to-sameorigin-creds?url=redirect.py%3FRedirect%3Dhttps%253A%252F%252Ffoo%253Abar%2540localhost%253A9443%252Fservice-workers%252Fservice-worker%252Fresources%252Fsuccess.py Promise did not reject."
FAIL Non-navigation, error redirect, cors mode Request redirected to no-cors with credentials should fail interception and response should not be redirected promise_test: Unhandled rejection with value: "assert_promise_rejects: Must fail to fetch: url=/nonav-error-cors-redirects-to-nocors-creds?url=redirect.py%3FRedirect%3Dhttps%253A%252F%252Ffoo%253Abar%2540127.0.0.1%253A9443%252Fservice-workers%252Fservice-worker%252Fresources%252Fsuccess.py Promise did not reject."
FAIL Non-navigation, error redirect, cors mode Request redirected to cors with credentials should fail interception and response should not be redirected promise_test: Unhandled rejection with value: "assert_promise_rejects: Must fail to fetch: url=/nonav-error-cors-redirects-to-cors-creds?url=redirect.py%3FRedirect%3Dhttps%253A%252F%252Ffoo%253Abar%2540127.0.0.1%253A9443%252Fservice-workers%252Fservice-worker%252Fresources%252Fsuccess.py%253FACAOrigin%253Dhttps%25253A%25252F%25252Flocalhost%25253A9443 Promise did not reject."
FAIL Non-navigation, error redirect, same-origin mode Request redirected to same-origin with credentials should fail interception and response should not be redirected promise_test: Unhandled rejection with value: "assert_promise_rejects: Must fail to fetch: url=/nonav-error-sameorigin-redirects-to-sameorigin-creds?url=redirect.py%3FRedirect%3Dhttps%253A%252F%252Ffoo%253Abar%2540localhost%253A9443%252Fservice-workers%252Fservice-worker%252Fresources%252Fsuccess.py Promise did not reject."
FAIL Non-navigation, error redirect, same-origin mode Request redirected to no-cors with credentials should fail interception and response should not be redirected promise_test: Unhandled rejection with value: "assert_promise_rejects: Must fail to fetch: url=/nonav-error-sameorigin-redirects-to-nocors-creds?url=redirect.py%3FRedirect%3Dhttps%253A%252F%252Ffoo%253Abar%2540127.0.0.1%253A9443%252Fservice-workers%252Fservice-worker%252Fresources%252Fsuccess.py Promise did not reject."
FAIL Non-navigation, error redirect, same-origin mode Request redirected to cors with credentials should fail interception and response should not be redirected promise_test: Unhandled rejection with value: "assert_promise_rejects: Must fail to fetch: url=/nonav-error-sameorigin-redirects-to-cors-creds?url=redirect.py%3FRedirect%3Dhttps%253A%252F%252Ffoo%253Abar%2540127.0.0.1%253A9443%252Fservice-workers%252Fservice-worker%252Fresources%252Fsuccess.py%253FACAOrigin%253Dhttps%25253A%25252F%25252Flocalhost%25253A9443 Promise did not reject."
FAIL Non-navigation, error redirect, no-cors mode Request redirected to same-origin with credentials should fail interception and response should not be redirected promise_test: Unhandled rejection with value: "assert_promise_rejects: Must fail to fetch: url=/nonav-error-nocors-redirects-to-sameorigin-creds?url=redirect.py%3FRedirect%3Dhttps%253A%252F%252Ffoo%253Abar%2540localhost%253A9443%252Fservice-workers%252Fservice-worker%252Fresources%252Fsuccess.py Promise did not reject."
FAIL Non-navigation, error redirect, no-cors mode Request redirected to no-cors with credentials should fail interception and response should not be redirected promise_test: Unhandled rejection with value: "assert_promise_rejects: Must fail to fetch: url=/nonav-error-nocors-redirects-to-nocors-creds?url=redirect.py%3FRedirect%3Dhttps%253A%252F%252Ffoo%253Abar%2540127.0.0.1%253A9443%252Fservice-workers%252Fservice-worker%252Fresources%252Fsuccess.py Promise did not reject."
FAIL Non-navigation, error redirect, no-cors mode Request redirected to cors with credentials should fail interception and response should not be redirected promise_test: Unhandled rejection with value: "assert_promise_rejects: Must fail to fetch: url=/nonav-error-nocors-redirects-to-cors-creds?url=redirect.py%3FRedirect%3Dhttps%253A%252F%252Ffoo%253Abar%2540127.0.0.1%253A9443%252Fservice-workers%252Fservice-worker%252Fresources%252Fsuccess.py%253FACAOrigin%253Dhttps%25253A%25252F%25252Flocalhost%25253A9443 Promise did not reject."
PASS restore global state 
